+++ LookupAnything/GameHelper.cs
@@ -240,12 +240,13 @@ internal class GameHelper
 
     /// <summary>Count how many of an item the player owns.</summary>
     /// <param name="item">The item to count.</param>
-    public int CountOwnedItems(Item item)
+    /// <param name="flavorSpecific">If the item has a preserved item, whether to only include items with the same preserved item.</param>
+    public int CountOwnedItems(Item item, bool flavorSpecific)
     {
         return (
             from found in this.GetAllOwnedItems()
             let foundItem = found.Item
-            where this.AreEquivalent(foundItem, item)
+            where this.AreEquivalent(foundItem, item, flavorSpecific)
             let canStack = foundItem.canStackWith(foundItem)
             select canStack ? found.GetCount() : 1
         ).Sum();
@@ -463,7 +464,7 @@ internal class GameHelper
     {
         return this
             .GetRecipes()
-            .Where(recipe => this.AreEquivalent(item, recipe.TryCreateItem(item)));
+            .Where(recipe => this.AreEquivalent(item, recipe.TryCreateItem(item), flavorSpecific: false));
     }
 
     /// <summary>Get the recipes for a given machine.</summary>
@@ -686,13 +687,18 @@ internal class GameHelper
     /// <summary>Get whether two items are the same type (ignoring flavor text like 'blueberry wine' vs 'cranberry wine').</summary>
     /// <param name="a">The first item to compare.</param>
     /// <param name="b">The second item to compare.</param>
-    private bool AreEquivalent(Item? a, Item? b)
+    /// <param name="flavorSpecific">Whether to only match items with the same preserved item.</param>
+    private bool AreEquivalent(Item? a, Item? b, bool flavorSpecific)
     {
         return
             a != null
             && b != null
             && a.QualifiedItemId == b.QualifiedItemId
-            && (a as Chest)?.fridge.Value == (b as Chest)?.fridge.Value;
+            && (a as Chest)?.fridge.Value == (b as Chest)?.fridge.Value
+            && (
+                !flavorSpecific
+                || (a as SObject)?.GetPreservedItemId() == (b as SObject)?.GetPreservedItemId()
+            );
     }
 
     /// <summary>Get all machine recipes, including those from mods like Producer Framework Mod.</summary>